Improve UI + fix some issues
